requirements.txt
playwright
beautifulsoup4
requests
anthropic
fal-client
python-dotenv
boto3
psycopg2-binary
